loaders: simplify source loader interceptors
